Fix compiler warning
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@1059 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/fbutil/FirebirdConfigFile.cpp
@@ -18,7 +18,7 @@
 
 void FirebirdConfigFile::key_compare::uppercase(string& x) const
 {
-	for (int pos = 0; pos < x.length(); pos++)
+	for (size_t pos = 0; pos < x.length(); pos++)
 		x[pos] = toupper(x[pos]);
 }
